Fix Go code review findings across all packages
Addresses 24 findings from a full Go code review, spanning bug fixes,
robustness, maintainability, and style.

Bug risk (P0):
- gitquery: mark a worktree stale on any os.Stat failure, not just
  os.IsNotExist (catches permission errors and other access failures)
- model: surface force-delete failures via ForceDeleteFailedMsg instead
  of silently reporting success

Robustness (P1):
- model: data-loading/diff commands now return FetchErrorMsg (and stash
  drop / worktree prune return ActionFailedMsg) instead of swallowing
  errors and rendering blank panes
- actions: wrap CreateWorktree's error with %w; route DeleteBranch,
  ForceDeleteBranch, and DropStash through runGit so failures surface the
  trimmed git stderr rather than "exit status 1"
- actions: validate $SHELL (exec-bit / LookPath) before exec, fall back
  to /bin/sh
- scanner: recognize .git as a regular file (worktree pointers), not only
  a directory

Maintainability (P2):
- model: extract resetModeCursors, moveItemCursor/moveCursor, and merge
  fetchTargetPath/pullTargetPath into gitTargetPath(forPull)
- model: split the 2000-line model.go into model_keys.go, model_fetch.go,
  model_messages.go, and model_filter.go (no behavior change)
- gitquery: extract shared readDirtyStatus helper
- ui: replace renderStatusBarWithState's 14 positional params with a
  statusBarParams struct
- ui: move Mode type into the ui package and replace magic-number mode
  comparisons with named constants
- actions: remove test-only OpenTerminal; document scanner MaxDepth limits

Style (P3):
- actions: replace crypto/sha1 fingerprint with hash/fnv
- internal/version: unexport Version/Commit/Date (update Makefile and
  .goreleaser.yml ldflags paths accordingly)
- model: drop the OverlayState re-export, reference ui.OverlayState directly
- ui: move dead renderBranchPane into export_test.go; comment ANSI palette
- gitquery: remove a join/split round-trip in ParseWorktreeList

All packages build, vet, and test clean; gofmt clean.
