+++ model/model_action_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/brian-bell/wtui/gitquery"
 	"github.com/brian-bell/wtui/model"
+	"github.com/brian-bell/wtui/ui"
 )
 
 // --- Worktree diff (enter key in ModeWorktrees) ---
@@ -22,16 +23,15 @@ func TestModel_EnterOnDirtyWorktreeOpensDiffOverlay(t *testing.T) {
 	m, _ = update(m, model.WorktreeResultMsg{RepoPath: "/dev/alpha", Worktrees: wts})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayWorktreeDiff {
+	if m.Overlay() != ui.OverlayWorktreeDiff {
 		t.Errorf("expected OverlayWorktreeDiff, got %d", m.Overlay())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchWorktreeDiff cmd, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeDiffResultMsg); !ok {
-		t.Errorf("expected WorktreeDiffResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeDiffResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeDiffResultMsg")
 }
 
 func TestModel_EnterOnCleanWorktreeIsNoOp(t *testing.T) {
@@ -43,7 +43,7 @@ func TestModel_EnterOnCleanWorktreeIsNoOp(t *testing.T) {
 	m, _ = update(m, model.WorktreeResultMsg{RepoPath: "/dev/alpha", Worktrees: wts})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone for clean worktree, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -60,7 +60,7 @@ func TestModel_EnterOnStaleWorktreeIsNoOp(t *testing.T) {
 	m, _ = update(m, model.WorktreeResultMsg{RepoPath: "/dev/alpha", Worktrees: wts})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone for stale worktree, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -77,7 +77,7 @@ func TestModel_EnterOnLockedDirtyWorktreeOpensDiffOverlay(t *testing.T) {
 	m, _ = update(m, model.WorktreeResultMsg{RepoPath: "/dev/alpha", Worktrees: wts})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayWorktreeDiff {
+	if m.Overlay() != ui.OverlayWorktreeDiff {
 		t.Errorf("expected OverlayWorktreeDiff for locked dirty worktree, got %d", m.Overlay())
 	}
 	if cmd == nil {
@@ -90,7 +90,7 @@ func TestModel_EnterOnEmptyWorktreeListIsNoOp(t *testing.T) {
 	m = inRightPane(m)
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone with no worktrees, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -326,14 +326,14 @@ func TestModel_FAndShiftFKeys_NonWorktreeAndBranchModes_NoCmd(t *testing.T) {
 	for _, tc := range []struct {
 		name string
 		key  rune
-		mode model.Mode
+		mode ui.Mode
 	}{
-		{name: "fetch stashes", key: 'f', mode: model.ModeStashes},
-		{name: "pull stashes", key: 'F', mode: model.ModeStashes},
-		{name: "fetch history", key: 'f', mode: model.ModeHistory},
-		{name: "pull history", key: 'F', mode: model.ModeHistory},
-		{name: "fetch reflog", key: 'f', mode: model.ModeReflog},
-		{name: "pull reflog", key: 'F', mode: model.ModeReflog},
+		{name: "fetch stashes", key: 'f', mode: ui.ModeStashes},
+		{name: "pull stashes", key: 'F', mode: ui.ModeStashes},
+		{name: "fetch history", key: 'f', mode: ui.ModeHistory},
+		{name: "pull history", key: 'F', mode: ui.ModeHistory},
+		{name: "fetch reflog", key: 'f', mode: ui.ModeReflog},
+		{name: "pull reflog", key: 'F', mode: ui.ModeReflog},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			m := model.New(testRepos())
@@ -357,9 +357,8 @@ func TestModel_GitFetchedRefetchesCurrentMode(t *testing.T) {
 		t.Fatal("expected refetch cmd after fetch success")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.BranchResultMsg); !ok {
-		t.Fatalf("expected BranchResultMsg from refetch, got %T", msg)
-	}
+	_, ok := msg.(model.BranchResultMsg)
+	assertFetchMsg(t, msg, ok, "expected BranchResultMsg from refetch")
 }
 
 func TestModel_GitPulledRefetchesCurrentMode(t *testing.T) {
@@ -371,9 +370,8 @@ func TestModel_GitPulledRefetchesCurrentMode(t *testing.T) {
 		t.Fatal("expected refetch cmd after pull success")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Fatalf("expected WorktreeResultMsg from refetch, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeResultMsg from refetch")
 }
 
 func TestModel_StaleGitFetchedMsgIgnored(t *testing.T) {
@@ -415,10 +413,8 @@ func TestModel_EnterStillRequiresDirtyWorktree(t *testing.T) {
 	}
 	msg := cmd()
 	diffMsg, ok := msg.(model.BranchDiffResultMsg)
-	if !ok {
-		t.Fatalf("expected BranchDiffResultMsg, got %T", msg)
-	}
-	if diffMsg.BranchName != "dirty-root" {
+	assertFetchMsg(t, msg, ok, "BranchDiffResultMsg")
+	if ok && diffMsg.BranchName != "dirty-root" {
 		t.Errorf("expected dirty-root, got %q", diffMsg.BranchName)
 	}
 
@@ -452,16 +448,15 @@ func TestModel_EnterOpensBranchDiffOverlayForDirtyWorktree(t *testing.T) {
 	m, _ = update(m, model.BranchResultMsg{RepoPath: "/dev/alpha", Branches: branches})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayBranchDiff {
+	if m.Overlay() != ui.OverlayBranchDiff {
 		t.Errorf("expected OverlayBranchDiff, got %d", m.Overlay())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchBranchDiff cmd, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.BranchDiffResultMsg); !ok {
-		t.Errorf("expected BranchDiffResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.BranchDiffResultMsg)
+	assertFetchMsg(t, msg, ok, "expected BranchDiffResultMsg")
 }
 
 func TestModel_EnterDoesNothingForCleanBranch(t *testing.T) {
@@ -476,7 +471,7 @@ func TestModel_EnterDoesNothingForCleanBranch(t *testing.T) {
 	m, _ = update(m, model.BranchResultMsg{RepoPath: "/dev/alpha", Branches: branches})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -497,16 +492,15 @@ func modelInHistoryWithCommits() model.Model {
 func TestModel_EnterInHistoryOpensCommitDiffOverlay(t *testing.T) {
 	m := modelInHistoryWithCommits()
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayCommitDiff {
+	if m.Overlay() != ui.OverlayCommitDiff {
 		t.Errorf("expected OverlayCommitDiff, got %d", m.Overlay())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchCommitDiff cmd, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.CommitDiffResultMsg); !ok {
-		t.Errorf("expected CommitDiffResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.CommitDiffResultMsg)
+	assertFetchMsg(t, msg, ok, "expected CommitDiffResultMsg")
 }
 
 func TestModel_EnterInHistoryNoCommitsIsNoOp(t *testing.T) {
@@ -515,7 +509,7 @@ func TestModel_EnterInHistoryNoCommitsIsNoOp(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'4'}})
 	// No commits loaded
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -594,7 +588,7 @@ func TestModel_DKeyNoOpInHistoryMode(t *testing.T) {
 	m := modelInHistoryWithCommits()
 	m = enableDestructive(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone in history mode, got %d", m.Overlay())
 	}
 }
@@ -623,16 +617,15 @@ func TestModel_EnterOpensOverlay(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'3'}})
 	m, _ = update(m, model.StashResultMsg{RepoPath: "/dev/alpha", Stashes: testStashes()})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayStashDiff {
+	if m.Overlay() != ui.OverlayStashDiff {
 		t.Errorf("expected OverlayStashDiff, got %d", m.Overlay())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchStashDiff cmd, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.StashDiffResultMsg); !ok {
-		t.Errorf("expected StashDiffResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.StashDiffResultMsg)
+	assertFetchMsg(t, msg, ok, "expected StashDiffResultMsg")
 }
 
 func TestModel_StashDiffResultStoresDiff(t *testing.T) {
@@ -651,7 +644,7 @@ func TestModel_EscClosesOverlay(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyEnter})
 	// Now close overlay with esc
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEscape})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -670,7 +663,7 @@ func TestModel_QClosesOverlayNotQuit(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyEnter})
 	// Close with q
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'q'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -715,7 +708,7 @@ func TestModel_ModeKeysIgnoredInOverlay(t *testing.T) {
 	if m.Mode() != 3 {
 		t.Errorf("expected mode unchanged at 3 (stashes), got %d", m.Mode())
 	}
-	if m.Overlay() != model.OverlayStashDiff {
+	if m.Overlay() != ui.OverlayStashDiff {
 		t.Errorf("expected overlay still open, got %d", m.Overlay())
 	}
 }
@@ -731,7 +724,7 @@ func TestModel_DKeyNoOpInReadOnlyMode(t *testing.T) {
 	})
 	// d should be no-op in read-only mode (default)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone in read-only mode, got %d", m.Overlay())
 	}
 }
@@ -759,7 +752,7 @@ func TestModel_DKeyWorksInDestructiveMode(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'D'}})
 	// Now d should work
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("expected OverlayConfirm in destructive mode, got %d", m.Overlay())
 	}
 }
@@ -770,7 +763,7 @@ func TestModel_DKeyNoOpInReadOnlyModeStashes(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'3'}})
 	m, _ = update(m, model.StashResultMsg{RepoPath: "/dev/alpha", Stashes: testStashes()})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone for stash drop in read-only mode, got %d", m.Overlay())
 	}
 }
@@ -810,7 +803,7 @@ func TestModel_ShiftDNoOpDuringConfirmOverlay(t *testing.T) {
 	m := modelWithDeletableBranch()
 	// Open confirm dialog
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Fatal("expected OverlayConfirm")
 	}
 	// Shift+D should be ignored while confirm is active
@@ -831,7 +824,7 @@ func TestModel_ShiftDNoOpDuringDiffOverlay(t *testing.T) {
 	})
 	// Open diff overlay
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() == model.OverlayNone {
+	if m.Overlay() == ui.OverlayNone {
 		t.Fatal("expected a diff overlay")
 	}
 	// Not in destructive mode; Shift+D should be ignored
@@ -850,16 +843,15 @@ func TestModel_WorktreeRemovedDetachedSkipsBranchConfirm(t *testing.T) {
 	}})
 	// Send WorktreeRemovedMsg with empty BranchName (detached)
 	m, cmd := update(m, model.WorktreeRemovedMsg{RepoPath: "/dev/alpha", BranchName: ""})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("detached removal should not show branch confirm, got overlay %d", m.Overlay())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchWorktrees cmd after detached removal, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg from refetch, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeResultMsg from refetch")
 }
 
 func TestModel_WorktreeRemovedShowsBranchConfirm(t *testing.T) {
@@ -871,7 +863,7 @@ func TestModel_WorktreeRemovedShowsBranchConfirm(t *testing.T) {
 	}})
 	// Send WorktreeRemovedMsg with branch name
 	m, cmd := update(m, model.WorktreeRemovedMsg{RepoPath: "/dev/alpha", BranchName: "feat"})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("non-detached removal should show branch confirm, got overlay %d", m.Overlay())
 	}
 	if !strings.Contains(m.ConfirmPrompt(), "feat") {
@@ -885,9 +877,8 @@ func TestModel_WorktreeRemovedShowsBranchConfirm(t *testing.T) {
 		t.Fatal("expected fetchWorktrees cmd alongside branch confirm, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg from background refetch, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeResultMsg from background refetch")
 }
 
 func TestModel_CombinedCleanupConfirmYReturnsCmd(t *testing.T) {
@@ -899,12 +890,12 @@ func TestModel_CombinedCleanupConfirmYReturnsCmd(t *testing.T) {
 	}})
 	// Trigger branch confirm
 	m, _ = update(m, model.WorktreeRemovedMsg{RepoPath: "/dev/alpha", BranchName: "feat"})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Fatalf("expected OverlayConfirm, got %d", m.Overlay())
 	}
 	// Confirm branch deletion
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'y'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed after confirm, got %d", m.Overlay())
 	}
 	if cmd == nil {
@@ -927,7 +918,7 @@ func TestModel_CombinedCleanupConfirmNClosesDialog(t *testing.T) {
 	m, _ = update(m, model.WorktreeRemovedMsg{RepoPath: "/dev/alpha", BranchName: "feat"})
 	// Decline branch deletion
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed after cancel, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -947,7 +938,7 @@ func TestModel_CombinedCleanupForceDeleteReturnsWorktreeDeleteCompletedMsg(t *te
 	}})
 	// Worktree removed → branch confirm dialog
 	m, _ = update(m, model.WorktreeRemovedMsg{RepoPath: "/dev/alpha", BranchName: "feat"})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Fatalf("expected branch confirm overlay, got %d", m.Overlay())
 	}
 	// Confirm branch deletion → DeleteBranch fails on fake path → DeleteFailedMsg
@@ -961,7 +952,7 @@ func TestModel_CombinedCleanupForceDeleteReturnsWorktreeDeleteCompletedMsg(t *te
 	}
 	// Process DeleteFailedMsg → force confirm shown
 	m, _ = update(m, deleteFailedMsg)
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Fatalf("expected force confirm overlay, got %d", m.Overlay())
 	}
 	if !m.ConfirmForce() {
@@ -973,9 +964,8 @@ func TestModel_CombinedCleanupForceDeleteReturnsWorktreeDeleteCompletedMsg(t *te
 		t.Fatal("expected force cmd, got nil")
 	}
 	result := forceCmd()
-	if _, ok := result.(model.WorktreeDeleteCompletedMsg); !ok {
-		t.Errorf("force-delete in combined cleanup should return WorktreeDeleteCompletedMsg, got %T", result)
-	}
+	_, ok := result.(model.WorktreeDeleteCompletedMsg)
+	assertFetchMsg(t, result, ok, "force-delete in combined cleanup should return WorktreeDeleteCompletedMsg")
 }
 
 // --- Worktree prune ---
@@ -988,7 +978,7 @@ func TestModel_PKeyRequiresDestructiveMode(t *testing.T) {
 		{Path: "/dev/gone", BranchName: "stale", Stale: true},
 	}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'p'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("p without destructive mode should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -998,7 +988,7 @@ func TestModel_PKeyNoOpOnNonStaleWorktree(t *testing.T) {
 		{Path: "/dev/alpha-feat", BranchName: "feat"},
 	})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'p'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("p on non-stale worktree should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -1008,7 +998,7 @@ func TestModel_PKeyOnStaleWorktreeShowsConfirm(t *testing.T) {
 		{Path: "/dev/gone", BranchName: "stale-branch", Stale: true},
 	})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'p'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("p on stale worktree should open confirm, got overlay %d", m.Overlay())
 	}
 	if !strings.Contains(m.ConfirmPrompt(), "Prune") {
@@ -1021,7 +1011,7 @@ func TestModel_PKeyNoOpOnLockedStaleWorktree(t *testing.T) {
 		{Path: "/dev/gone", BranchName: "offline", Locked: true, Stale: true},
 	})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'p'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("p on locked stale worktree should be no-op, got overlay %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -1034,7 +1024,7 @@ func TestModel_PKeyNoOpInBranchesMode(t *testing.T) {
 	m = inBranchesMode(m)
 	m = enableDestructive(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'p'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("p in branches mode should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -1051,9 +1041,8 @@ func TestModel_WorktreePrunedRefetchesWorktrees(t *testing.T) {
 		t.Fatal("expected fetchWorktrees cmd after prune, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg from refetch, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeResultMsg from refetch")
 }
 
 // --- Worktree t/c actions ---
@@ -1097,7 +1086,7 @@ func TestModel_TKeyOnStaleWorktreeIsNoOp(t *testing.T) {
 func TestModel_WorktreeDeleteCompletedMsgIsNoOp(t *testing.T) {
 	m := model.New(testRepos())
 	m, cmd := update(m, model.WorktreeDeleteCompletedMsg{RepoPath: "/dev/alpha"})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -1111,7 +1100,7 @@ func TestModel_NKeyOpensWorktreeInput(t *testing.T) {
 	m := model.New(testRepos())
 	m = inWorktreesMode(m)
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
-	if m.Overlay() != model.OverlayWorktreeInput {
+	if m.Overlay() != ui.OverlayWorktreeInput {
 		t.Errorf("expected OverlayWorktreeInput, got %d", m.Overlay())
 	}
 	if m.WorktreeInput() != "" {
@@ -1126,12 +1115,12 @@ func TestModel_NKeyNoOpOutsideWorktreesMode(t *testing.T) {
 	for _, tc := range []struct {
 		name string
 		key  rune
-		mode model.Mode
+		mode ui.Mode
 	}{
-		{name: "branches", key: '2', mode: model.ModeBranches},
-		{name: "stashes", key: '3', mode: model.ModeStashes},
-		{name: "history", key: '4', mode: model.ModeHistory},
-		{name: "reflog", key: '5', mode: model.ModeReflog},
+		{name: "branches", key: '2', mode: ui.ModeBranches},
+		{name: "stashes", key: '3', mode: ui.ModeStashes},
+		{name: "history", key: '4', mode: ui.ModeHistory},
+		{name: "reflog", key: '5', mode: ui.ModeReflog},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			m := model.New(testRepos())
@@ -1142,7 +1131,7 @@ func TestModel_NKeyNoOpOutsideWorktreesMode(t *testing.T) {
 			}
 
 			m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
-			if m.Overlay() != model.OverlayNone {
+			if m.Overlay() != ui.OverlayNone {
 				t.Errorf("expected OverlayNone, got %d", m.Overlay())
 			}
 			if cmd != nil {
@@ -1169,7 +1158,7 @@ func TestModel_WorktreeInputEscCancels(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune("feat")})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEscape})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed, got %d", m.Overlay())
 	}
 	if m.WorktreeInput() != "" {
@@ -1186,7 +1175,7 @@ func TestModel_WorktreeInputCtrlCCancels(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune("feat")})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyCtrlC})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed, got %d", m.Overlay())
 	}
 	if m.WorktreeInput() != "" {
@@ -1202,7 +1191,7 @@ func TestModel_WorktreeInputEnterRequiresText(t *testing.T) {
 	m = inWorktreesMode(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayWorktreeInput {
+	if m.Overlay() != ui.OverlayWorktreeInput {
 		t.Errorf("expected input overlay to remain, got %d", m.Overlay())
 	}
 	if m.WorktreeInputErr() == "" {
@@ -1219,7 +1208,7 @@ func TestModel_WorktreeInputEnterCreatesWorktree(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune("feat")})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed, got %d", m.Overlay())
 	}
 	if cmd == nil {
@@ -1235,22 +1224,21 @@ func TestModel_WorktreeCreatedRefetchesWorktrees(t *testing.T) {
 	m := model.New(testRepos())
 	m = inBranchesMode(m)
 	m, cmd := update(m, model.WorktreeCreatedMsg{RepoPath: "/dev/alpha", WorktreePath: "/dev/alpha-worktrees/feat"})
-	if m.Mode() != model.ModeWorktrees {
+	if m.Mode() != ui.ModeWorktrees {
 		t.Errorf("expected mode worktrees after create, got %d", m.Mode())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchWorktrees cmd after create")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg from refetch, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeResultMsg from refetch")
 }
 
 func TestModel_WorktreeCreateFailedReopensInput(t *testing.T) {
 	m := model.New(testRepos())
 	m, _ = update(m, model.WorktreeCreateFailedMsg{RepoPath: "/dev/alpha", Input: "feat", Err: "boom"})
-	if m.Overlay() != model.OverlayWorktreeInput {
+	if m.Overlay() != ui.OverlayWorktreeInput {
 		t.Errorf("expected OverlayWorktreeInput, got %d", m.Overlay())
 	}
 	if m.WorktreeInput() != "feat" {
@@ -1264,7 +1252,7 @@ func TestModel_WorktreeCreateFailedReopensInput(t *testing.T) {
 func TestModel_WorktreeCreateFailedUsesFallbackError(t *testing.T) {
 	m := model.New(testRepos())
 	m, _ = update(m, model.WorktreeCreateFailedMsg{RepoPath: "/dev/alpha", Input: "feat"})
-	if m.Overlay() != model.OverlayWorktreeInput {
+	if m.Overlay() != ui.OverlayWorktreeInput {
 		t.Errorf("expected OverlayWorktreeInput, got %d", m.Overlay())
 	}
 	if m.WorktreeInput() != "feat" {
@@ -1297,7 +1285,7 @@ func modelWithDeletableBranch() model.Model {
 func TestModel_DKeyOpensConfirmOverlay(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("expected OverlayConfirm, got %d", m.Overlay())
 	}
 	if !strings.Contains(m.ConfirmPrompt(), "feat") {
@@ -1314,7 +1302,7 @@ func TestModel_DKeyOnNonWorktreeBranchOpensDeleteConfirm(t *testing.T) {
 	})
 	m = enableDestructive(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("expected OverlayConfirm for non-worktree branch, got %d", m.Overlay())
 	}
 	if !strings.Contains(m.ConfirmPrompt(), "main") {
@@ -1330,7 +1318,7 @@ func TestModel_DKeyNoOpWithNoBranches(t *testing.T) {
 	// No branches loaded
 	m = inRightPane(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone when no branches, got %d", m.Overlay())
 	}
 }
@@ -1339,7 +1327,7 @@ func TestModel_ConfirmCancelEsc(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEscape})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed on esc, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -1354,7 +1342,7 @@ func TestModel_ConfirmCancelQ(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'q'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed on q, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -1369,7 +1357,7 @@ func TestModel_ConfirmCancelN(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed on n, got %d", m.Overlay())
 	}
 }
@@ -1378,7 +1366,7 @@ func TestModel_ConfirmYClosesOverlayAndReturnsCmd(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'y'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed after confirm, got %d", m.Overlay())
 	}
 	if cmd == nil {
@@ -1390,7 +1378,7 @@ func TestModel_ConfirmEnterExecutesAction(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed after enter, got %d", m.Overlay())
 	}
 	if cmd == nil {
@@ -1429,7 +1417,7 @@ func TestModel_DeleteFailedMsgOpensForceConfirm(t *testing.T) {
 			return nil
 		},
 	})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("expected OverlayConfirm after DeleteFailedMsg, got %d", m.Overlay())
 	}
 	if !m.ConfirmForce() {
@@ -1452,7 +1440,7 @@ func TestModel_ForceConfirmCancelClearsForce(t *testing.T) {
 		ForceAction: func() error { return nil },
 	})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'n'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed after cancel, got %d", m.Overlay())
 	}
 	if m.ConfirmForce() {
@@ -1468,7 +1456,7 @@ func TestModel_ForceConfirmYExecutesForceAction(t *testing.T) {
 		ForceAction: func() error { return nil },
 	})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'y'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected overlay closed after force confirm, got %d", m.Overlay())
 	}
 	if m.ConfirmForce() {
@@ -1487,7 +1475,7 @@ func TestModel_ConfirmDialogBlocksModeSwitch(t *testing.T) {
 	m := modelWithDeletableBranch()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'3'}})
-	if m.Mode() != model.ModeBranches {
+	if m.Mode() != ui.ModeBranches {
 		t.Errorf("confirm dialog should block mode switch, mode changed to %d", m.Mode())
 	}
 }
@@ -1506,7 +1494,7 @@ func modelInStashesWithStashes() model.Model {
 func TestModel_DKeyInStashesModeOpensConfirmDialog(t *testing.T) {
 	m := modelInStashesWithStashes()
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("expected OverlayConfirm, got %d", m.Overlay())
 	}
 	if !strings.Contains(m.ConfirmPrompt(), "stash@{0}") {
@@ -1520,7 +1508,7 @@ func TestModel_DKeyInStashesModeWithNoStashesDoesNothing(t *testing.T) {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'3'}})
 	// No stashes loaded
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone when no stashes, got %d", m.Overlay())
 	}
 }
@@ -1533,9 +1521,8 @@ func TestModel_StashDropConfirmReturnsStashDroppedMsg(t *testing.T) {
 		t.Fatal("expected cmd after stash drop confirm, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.StashDroppedMsg); !ok {
-		t.Errorf("expected StashDroppedMsg, got %T", msg)
-	}
+	_, ok := msg.(model.StashDroppedMsg)
+	assertFetchMsg(t, msg, ok, "expected StashDroppedMsg")
 }
 
 // --- Open terminal / code ---
@@ -1614,14 +1601,14 @@ func TestModel_DKeyNoOpOnRootBranch(t *testing.T) {
 
 	// Cursor at root branch (pinned to index 0) — d should be no-op
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("d on root branch should be no-op, got overlay %d", m.Overlay())
 	}
 
 	// Navigate to feat (index 1) — d should open confirm
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyDown})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("d on non-root branch should open confirm, got overlay %d", m.Overlay())
 	}
 }
@@ -1648,7 +1635,7 @@ func TestModel_DKeyNoOpOnRootWorktree(t *testing.T) {
 	})
 	// Cursor at root worktree (index 0) — d should be no-op
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("d on root worktree should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -1661,7 +1648,7 @@ func TestModel_DKeyNoOpOnStaleWorktree(t *testing.T) {
 	// Navigate to stale worktree (index 1)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyDown})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("d on stale worktree should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -1673,7 +1660,7 @@ func TestModel_DKeyNoOpOnLockedWorktree(t *testing.T) {
 	})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyDown})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("d on locked worktree should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -1688,7 +1675,7 @@ func TestModel_DKeyOnWorktreeRequiresDestructiveMode(t *testing.T) {
 	}})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyDown})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("d without destructive mode should be no-op, got overlay %d", m.Overlay())
 	}
 }
@@ -1738,7 +1725,7 @@ func TestModel_DKeyOnWorktreeShowsConfirm(t *testing.T) {
 	// Navigate to non-root worktree (index 1)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyDown})
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayConfirm {
+	if m.Overlay() != ui.OverlayConfirm {
 		t.Errorf("d on non-root worktree should open confirm, got overlay %d", m.Overlay())
 	}
 	if !strings.Contains(m.ConfirmPrompt(), "/dev/alpha-feat") {
@@ -1754,7 +1741,7 @@ func TestModel_UKeyOnLockedWorktreeFiresUnlockCmd(t *testing.T) {
 	}})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'u'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("u should not open an overlay, got %d", m.Overlay())
 	}
 	if cmd == nil {
@@ -1787,7 +1774,7 @@ func TestModel_UKeyOnUnlockedWorktreeIsNoOp(t *testing.T) {
 	}})
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'u'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("u on unlocked worktree should not open overlay, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -1800,7 +1787,7 @@ func TestModel_UKeyOutsideWorktreesModeIsNoOp(t *testing.T) {
 	m = inBranchesMode(m)
 
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'u'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("u outside worktrees mode should not open overlay, got %d", m.Overlay())
 	}
 	if cmd != nil {
@@ -1830,9 +1817,8 @@ func TestModel_WorktreeUnlockedMsgRefetchesWorktrees(t *testing.T) {
 		t.Fatal("expected fetchWorktrees cmd after unlock")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg from refetch, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "expected WorktreeResultMsg from refetch")
 }
 
 func TestModel_StaleWorktreeUnlockedMsgIgnored(t *testing.T) {
@@ -1859,7 +1845,7 @@ func TestModel_DKeyNoOpInReflogMode(t *testing.T) {
 	m := modelInReflogWithEntries()
 	m = enableDestructive(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'d'}})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone in reflog mode, got %d", m.Overlay())
 	}
 }
@@ -1885,16 +1871,15 @@ func TestModel_YKeyNoOpWithNoReflogs(t *testing.T) {
 func TestModel_EnterInReflogOpensReflogDiffOverlay(t *testing.T) {
 	m := modelInReflogWithEntries()
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayReflogDiff {
+	if m.Overlay() != ui.OverlayReflogDiff {
 		t.Errorf("expected OverlayReflogDiff, got %d", m.Overlay())
 	}
 	if cmd == nil {
 		t.Fatal("expected fetchReflogDiff cmd, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.ReflogDiffResultMsg); !ok {
-		t.Errorf("expected ReflogDiffResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.ReflogDiffResultMsg)
+	assertFetchMsg(t, msg, ok, "expected ReflogDiffResultMsg")
 }
 
 func TestModel_ReflogDiffResultStoresDiff(t *testing.T) {
@@ -1943,7 +1928,7 @@ func TestModel_EnterInReflogNoEntriesIsNoOp(t *testing.T) {
 	m = inRightPane(m)
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'5'}})
 	m, cmd := update(m, tea.KeyMsg{Type: tea.KeyEnter})
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Errorf("expected OverlayNone, got %d", m.Overlay())
 	}
 	if cmd != nil {
