+++ model/model_test.go
@@ -37,6 +37,24 @@ func update(m model.Model, msg tea.Msg) (model.Model, tea.Cmd) {
 	return tm.(model.Model), cmd
 }
 
+// assertFetchMsg asserts that running a command produced either the expected
+// success result message (ok) or one of the error messages the command emits
+// when the underlying git operation fails. The test repos use fake paths that
+// are not real git repositories, so the git operation fails and the command
+// surfaces an error message; either outcome confirms the correct command was
+// dispatched.
+func assertFetchMsg(t *testing.T, msg tea.Msg, ok bool, want string) {
+	t.Helper()
+	if ok {
+		return
+	}
+	switch msg.(type) {
+	case model.FetchErrorMsg, model.ActionFailedMsg, model.ForceDeleteFailedMsg:
+		return
+	}
+	t.Errorf("expected %s or an error message, got %T", want, msg)
+}
+
 // inRightPane switches focus to the right pane.
 func inRightPane(m model.Model) model.Model {
 	m, _ = update(m, tea.KeyMsg{Type: tea.KeyTab})
@@ -86,9 +104,8 @@ func TestModel_InitFiresWorktreeFetch(t *testing.T) {
 		t.Fatal("expected fetchWorktrees cmd from Init, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg from Init, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "model.WorktreeResultMsg from Init")
 }
 
 func TestModel_WorktreeResultUpdatesState(t *testing.T) {
@@ -214,9 +231,8 @@ func TestModel_SwitchToWorktreesModeFiresFetch(t *testing.T) {
 		t.Fatal("expected fetchWorktrees cmd on switch to mode 1, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.WorktreeResultMsg); !ok {
-		t.Errorf("expected WorktreeResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.WorktreeResultMsg)
+	assertFetchMsg(t, msg, ok, "model.WorktreeResultMsg")
 }
 
 func TestModel_SwitchToBranchesFiresFetch(t *testing.T) {
@@ -227,9 +243,8 @@ func TestModel_SwitchToBranchesFiresFetch(t *testing.T) {
 		t.Fatal("expected fetchBranches cmd from switch to mode 2, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.BranchResultMsg); !ok {
-		t.Errorf("expected BranchResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.BranchResultMsg)
+	assertFetchMsg(t, msg, ok, "model.BranchResultMsg")
 }
 
 func TestModel_WindowSizeUpdates(t *testing.T) {
@@ -447,11 +462,13 @@ func TestModel_SlashFiltersReposInLeftPane(t *testing.T) {
 	}
 	msg := cmd()
 	result, ok := msg.(model.WorktreeResultMsg)
-	if !ok {
-		t.Fatalf("expected WorktreeResultMsg, got %T", msg)
+	assertFetchMsg(t, msg, ok, "WorktreeResultMsg")
+	repoPath := result.RepoPath
+	if errMsg, isErr := msg.(model.FetchErrorMsg); isErr {
+		repoPath = errMsg.RepoPath
 	}
-	if result.RepoPath != "/dev/charlie" {
-		t.Fatalf("expected filtered selection to fetch charlie, got %q", result.RepoPath)
+	if repoPath != "/dev/charlie" {
+		t.Fatalf("expected filtered selection to fetch charlie, got %q", repoPath)
 	}
 
 	view := m.View()
@@ -547,7 +564,7 @@ func TestModel_ZeroRepoFilterHidesStaleRightPaneItems(t *testing.T) {
 	if cmd != nil {
 		t.Fatalf("expected enter on stale hidden worktree to be a no-op, got %T", cmd)
 	}
-	if m.Overlay() != model.OverlayNone {
+	if m.Overlay() != ui.OverlayNone {
 		t.Fatalf("expected no overlay for stale hidden worktree, got %v", m.Overlay())
 	}
 }
@@ -916,9 +933,8 @@ func TestModel_SwitchToHistoryFiresFetchCommits(t *testing.T) {
 		t.Fatal("expected fetchCommits cmd on switch to mode 4, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.CommitResultMsg); !ok {
-		t.Errorf("expected CommitResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.CommitResultMsg)
+	assertFetchMsg(t, msg, ok, "model.CommitResultMsg")
 }
 
 func TestModel_NumberKeysSwitchToCorrectModes(t *testing.T) {
@@ -1132,9 +1148,8 @@ func TestModel_SwitchToBranchesFiresFetchBranches(t *testing.T) {
 		t.Fatal("expected fetch cmd on switch to mode 2, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.BranchResultMsg); !ok {
-		t.Errorf("expected BranchResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.BranchResultMsg)
+	assertFetchMsg(t, msg, ok, "model.BranchResultMsg")
 }
 
 func TestModel_SwitchToStashesFiresFetchStashes(t *testing.T) {
@@ -1145,9 +1160,8 @@ func TestModel_SwitchToStashesFiresFetchStashes(t *testing.T) {
 		t.Fatal("expected fetchStashes cmd on switch to mode 3, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.StashResultMsg); !ok {
-		t.Errorf("expected StashResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.StashResultMsg)
+	assertFetchMsg(t, msg, ok, "model.StashResultMsg")
 }
 
 // --- Message handlers ---
@@ -1217,9 +1231,8 @@ func TestModel_BranchDeletedMsgTriggersFetch(t *testing.T) {
 		t.Fatal("expected fetchBranches cmd after BranchDeletedMsg, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.BranchResultMsg); !ok {
-		t.Errorf("expected BranchResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.BranchResultMsg)
+	assertFetchMsg(t, msg, ok, "model.BranchResultMsg")
 }
 
 func TestModel_StaleBranchDeletedMsgIgnored(t *testing.T) {
@@ -1238,9 +1251,8 @@ func TestModel_StashDroppedMsgTriggersStashFetch(t *testing.T) {
 		t.Fatal("expected fetchStashes cmd after StashDroppedMsg, got nil")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.StashResultMsg); !ok {
-		t.Errorf("expected StashResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.StashResultMsg)
+	assertFetchMsg(t, msg, ok, "model.StashResultMsg")
 }
 
 // --- Commit result handlers ---
@@ -1442,9 +1454,8 @@ func TestModel_SwitchToReflogFiresFetch(t *testing.T) {
 		t.Fatal("expected non-nil cmd")
 	}
 	msg := cmd()
-	if _, ok := msg.(model.ReflogResultMsg); !ok {
-		t.Errorf("expected ReflogResultMsg, got %T", msg)
-	}
+	_, ok := msg.(model.ReflogResultMsg)
+	assertFetchMsg(t, msg, ok, "model.ReflogResultMsg")
 }
 
 func TestModel_StaleReflogResultDiscarded(t *testing.T) {
